[WIP] indicator chooser migrated
